Fix error where we run out of bounds
Use an intermediate variable to prevent the index from running out of
bounds.

We not use i_cnt to count the number of times we've iterated and i as
out index which will only be updated when we need to calculate another
line.

Also move the check state block to better reflect the program flow.