Use another approach to calculate the digit product
We write the bits into the _raw var, then we simply concatenate 0s to
the front it it and save that into _shift.  We can then use the sll
operation to shift it to the correct location (we have the real digit
product of out two digits) and save that into _final and give it to
the rc_adder.